+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 - Deprecated prop modifier variant `withSetter("assign")`, use `withSetter()` instead.
+- Changed to nanoid from uuid for model id generation.
 
 ## 0.60.2